+++ gateway/gateway.go
@@ -615,6 +615,53 @@ func (gateway *HandleT) beaconBatchHandler(w http.ResponseWriter, r *http.Reques
 	gateway.beaconHandler(w, r, "batch")
 }
 
+type clearQueueRequestPayload struct {
+	SourceID      string `json:"source_id"`
+	DestinationID string `json:"destination_id"`
+}
+
+func (gateway *HandleT) ClearHandler(w http.ResponseWriter, r *http.Request) {
+	pkgLogger.LogRequest(r)
+	var errorMessage string
+	defer func() {
+		if errorMessage != "" {
+			pkgLogger.Debug(errorMessage)
+			http.Error(w, response.GetStatus(errorMessage), 400)
+		}
+	}()
+
+	payload, _, err := gateway.getPayloadAndWriteKey(w, r)
+	if err != nil {
+		errorMessage = err.Error()
+		return
+	}
+
+	if !gjson.ValidBytes(payload) {
+		errorMessage = response.GetStatus(response.InvalidJSON)
+		return
+	}
+
+	var reqPayload clearQueueRequestPayload
+	err = json.Unmarshal(payload, &reqPayload)
+	if err != nil {
+		errorMessage = err.Error()
+		return
+	}
+
+	if reqPayload.SourceID == "" {
+		errorMessage = "Empty source id"
+		return
+	}
+
+	err = queuemanager.QueueManager.InsertOperation(payload)
+	if err != nil {
+		errorMessage = err.Error()
+		return
+	}
+
+	w.Write([]byte(`{"status": "OK"}`))
+}
+
 type pendingEventsRequestPayload struct {
 	SourceID      string `json:"source_id"`
 	DestinationID string `json:"destination_id"`
@@ -1008,7 +1055,7 @@ func (gateway *HandleT) StartWebHandler() {
 	}
 
 	srvMux.HandleFunc("/v1/pending-events", gateway.stat(gateway.pendingEventsHandler))
-	srvMux.HandleFunc("/v1/clear", gateway.stat(queuemanager.ClearHandler))
+	srvMux.HandleFunc("/v1/clear", gateway.stat(gateway.ClearHandler))
 
 	c := cors.New(cors.Options{
 		AllowOriginFunc:  reflectOrigin,
